Fix mocking
