+++ internal/microsoft/content_azure_openai_text.go
@@ -13,7 +13,6 @@ import (
 	"github.com/hashicorp/hcl/v2"
 	"github.com/zclconf/go-cty/cty"
 
-	"github.com/blackstork-io/fabric/internal/microsoft/client"
 	"github.com/blackstork-io/fabric/pkg/diagnostics"
 	"github.com/blackstork-io/fabric/plugin"
 	"github.com/blackstork-io/fabric/plugin/dataspec"
@@ -21,7 +20,7 @@ import (
 	"github.com/blackstork-io/fabric/plugin/plugindata"
 )
 
-func makeAzureOpenAITextContentSchema(loader client.AzureOpenAIClientLoadFn) *plugin.ContentProvider {
+func makeAzureOpenAITextContentSchema(loader AzureOpenAIClientLoadFn) *plugin.ContentProvider {
 	return &plugin.ContentProvider{
 		Config: &dataspec.RootSpec{
 			Attrs: []*dataspec.AttrSpec{
@@ -81,7 +80,7 @@ func makeAzureOpenAITextContentSchema(loader client.AzureOpenAIClientLoadFn) *pl
 	}
 }
 
-func genOpenAIText(clientLoader client.AzureOpenAIClientLoadFn) plugin.ProvideContentFunc {
+func genOpenAIText(clientLoader AzureOpenAIClientLoadFn) plugin.ProvideContentFunc {
 	return func(ctx context.Context, params *plugin.ProvideContentParams) (*plugin.ContentResult, diagnostics.Diag) {
 		apiKey := params.Config.GetAttrVal("api_key").AsString()
 		resourceEndpoint := params.Config.GetAttrVal("resource_endpoint").AsString()
@@ -107,7 +106,7 @@ func genOpenAIText(clientLoader client.AzureOpenAIClientLoadFn) plugin.ProvideCo
 	}
 }
 
-func renderText(ctx context.Context, cli client.AzureOpenAIClient, cfg, args *dataspec.Block, dataCtx plugindata.Map) (string, error) {
+func renderText(ctx context.Context, cli AzureOpenAIClient, cfg, args *dataspec.Block, dataCtx plugindata.Map) (string, error) {
 	params := azopenai.CompletionsOptions{}
 	params.DeploymentName = to.Ptr(cfg.GetAttrVal("deployment_name").AsString())
